requirements.txt
Flask
gunicorn
Jinja2
keras
numpy
gunicorn
validators
scikit-learn
Werkzeug
itsdangerous
MarkupSafe
tensorflow-cpu